Prepare for a second implementation.
